Move tests related to runner switches to test_runner.py
