+++ tests/test_runner.py
@@ -6,16 +6,232 @@ import re
 import os
 import platform
 import subprocess
+import yaml
 
 from contextlib import redirect_stdout, redirect_stderr
 
 from lib.scenario_runner import ScenarioRunner
 from lib.global_config import GlobalConfig
 from lib.db import DB
+from lib import utils
 from lib.system_checks import ConfigurationCheckError
 from tests import test_functions as Tests
 
 GMT_DIR = os.path.join(os.path.dirname(os.path.abspath(__file__)), '../')
+
+### The Tests for the runner options/flags
+## --uri URI
+#   The URI to get the usage_scenario.yml from. Can be either a local directory starting with
+#     / or a remote git repository starting with http(s)://
+def test_uri_local_dir():
+    run_name = 'test_' + utils.randomword(12)
+    filename = 'tests/data/stress-application/usage_scenario.yml'
+    ps = subprocess.run(
+        ['python3', '../runner.py', '--name', run_name, '--uri', GMT_DIR ,'--config-override', f"{os.path.dirname(os.path.realpath(__file__))}/test-config.yml",
+        '--filename', filename,
+        '--skip-system-checks', '--dev-no-sleeps', '--dev-cache-build', '--dev-no-metrics', '--dev-no-phase-stats', '--dev-no-optimizations'],
+        check=True,
+        stderr=subprocess.PIPE,
+        stdout=subprocess.PIPE,
+        encoding='UTF-8'
+    )
+
+    filename_in_db = utils.get_run_data(run_name)['filename']
+    assert filename_in_db == filename, Tests.assertion_info(f"filename: {filename}", filename_in_db)
+    uri_in_db = utils.get_run_data(run_name)['uri']
+    assert uri_in_db == GMT_DIR, Tests.assertion_info(f"uri: {GMT_DIR}", uri_in_db)
+    assert ps.stderr == '', Tests.assertion_info('no errors', ps.stderr)
+
+def test_uri_local_dir_missing():
+    runner = ScenarioRunner(uri='/tmp/missing', uri_type='folder', filename='tests/data/usage_scenarios/basic_stress.yml', skip_system_checks=True, dev_no_metrics=True, dev_no_phase_stats=True, dev_no_sleeps=True, dev_cache_build=True)
+
+    with pytest.raises(FileNotFoundError) as e:
+        runner.run()
+
+
+    expected_exception = f"[Errno 2] No such file or directory: '{os.path.realpath('/tmp/missing')}'"
+
+    assert expected_exception == str(e.value),\
+        Tests.assertion_info(f"Exception: {expected_exception}", str(e.value))
+
+def test_non_git_root_supplied():
+    runner = ScenarioRunner(uri=f"{GMT_DIR}/tests/data/usage_scenarios/", uri_type='folder', filename='invalid_image.yml', skip_system_checks=True, dev_cache_build=False, dev_no_sleeps=True, dev_no_metrics=True, dev_no_phase_stats=True)
+
+    out = io.StringIO()
+    err = io.StringIO()
+    with redirect_stdout(out), redirect_stderr(err), pytest.raises(Exception) as e:
+        runner.run()
+
+    assert f"Supplied folder through --uri is not the root of the git repository. Please only supply the root folder and then the target directory through --filename. Real repo root is {GMT_DIR}" == str(e.value)
+
+    # basic positive case
+def test_uri_github_repo():
+    uri = 'https://github.com/green-coding-solutions/pytest-dummy-repo'
+    filename = 'usage_scenario.yml'
+    run_name = 'test_' + utils.randomword(12)
+    ps = subprocess.run(
+        ['python3', '../runner.py', '--name', run_name, '--uri', uri ,'--config-override', f"{os.path.dirname(os.path.realpath(__file__))}/test-config.yml",
+        '--skip-system-checks', '--dev-no-sleeps', '--dev-cache-build', '--dev-no-metrics', '--dev-no-phase-stats', '--dev-no-optimizations'],
+        check=True,
+        stderr=subprocess.PIPE,
+        stdout=subprocess.PIPE,
+        encoding='UTF-8'
+    )
+
+    filename_in_db = utils.get_run_data(run_name)['filename']
+    assert filename_in_db == filename, Tests.assertion_info(f"filename: {filename}", filename_in_db)
+    uri_in_db = utils.get_run_data(run_name)['uri']
+    assert uri_in_db == uri, Tests.assertion_info(f"uri: {uri}", uri_in_db)
+    assert ps.stderr == '', Tests.assertion_info('no errors', ps.stderr)
+
+## --branch BRANCH
+#    Optionally specify the git branch when targeting a git repository
+def test_uri_local_branch():
+    runner = ScenarioRunner(uri=GMT_DIR, uri_type='folder', filename='tests/data/usage_scenarios/basic_stress.yml', branch='test-branch', skip_system_checks=True, dev_no_metrics=True, dev_no_phase_stats=True, dev_no_sleeps=True, dev_cache_build=True)
+
+    out = io.StringIO()
+    err = io.StringIO()
+    with redirect_stdout(out), redirect_stderr(err), pytest.raises(RuntimeError) as e:
+        runner.run()
+    expected_exception = 'Specified --branch but using local URI. Did you mean to specify a github url?'
+    assert str(e.value) == expected_exception, \
+        Tests.assertion_info(f"Exception: {expected_exception}", str(e.value))
+
+    # basic positive case, branch prepped ahead of time
+    # this branch has a different usage_scenario file name - basic_stress
+    # that makes sure that it really is pulling a different branch
+def test_uri_github_repo_branch():
+    uri = 'https://github.com/green-coding-solutions/pytest-dummy-repo'
+    run_name = 'test_' + utils.randomword(12)
+    branch = 'test-branch'
+    ps = subprocess.run(
+        ['python3', '../runner.py', '--name', run_name, '--uri', uri ,
+        '--branch', branch , '--filename', 'basic_stress.yml',
+        '--config-override', f"{os.path.dirname(os.path.realpath(__file__))}/test-config.yml", '--skip-system-checks', '--dev-no-sleeps', '--dev-cache-build', '--dev-no-metrics', '--dev-no-phase-stats', '--dev-no-optimizations'],
+        check=True,
+        stderr=subprocess.PIPE,
+        stdout=subprocess.PIPE,
+        encoding='UTF-8'
+    )
+
+    branch_in_db = utils.get_run_data(run_name)['branch']
+    assert branch_in_db == 'test-branch', Tests.assertion_info(f"branch: {branch}", branch_in_db)
+    assert ps.stderr == '', Tests.assertion_info('no errors', ps.stderr)
+
+    # should throw error, assert vs error
+    # give incorrect branch name
+    ## Is the expected_exception OK or should it have a more graceful error?
+    ## ATM this is just the default console error of a failed git command
+def test_uri_github_repo_branch_missing():
+    runner = ScenarioRunner(uri='https://github.com/green-coding-solutions/pytest-dummy-repo', uri_type='URL', branch='missing-branch', filename='tests/data/usage_scenarios/basic_stress.yml', skip_system_checks=True, dev_no_metrics=True, dev_no_phase_stats=True, dev_no_sleeps=True, dev_cache_build=True)
+    with pytest.raises(subprocess.CalledProcessError) as e:
+        runner.run()
+    expected_exception = f"Command '['git', 'clone', '--depth', '1', '-b', 'missing-branch', '--single-branch', '--recurse-submodules', '--shallow-submodules', 'https://github.com/green-coding-solutions/pytest-dummy-repo', '{os.path.realpath('/tmp/green-metrics-tool/repo')}']' returned non-zero exit status 128."
+    assert expected_exception == str(e.value),\
+        Tests.assertion_info(f"Exception: {expected_exception}", str(e.value))
+
+# #   --name NAME
+# #    A name which will be stored to the database to discern this run from others
+def test_name_is_in_db():
+    run_name = 'test_' + utils.randomword(12)
+    subprocess.run(
+        ['python3', '../runner.py', '--name', run_name, '--uri', GMT_DIR ,
+        '--filename', 'tests/data/stress-application/usage_scenario.yml',
+        '--config-override', f"{os.path.dirname(os.path.realpath(__file__))}/test-config.yml",
+        '--skip-system-checks', '--dev-no-metrics', '--dev-no-phase-stats', '--dev-no-optimizations', '--dev-no-sleeps', '--dev-cache-build'],
+        check=True,
+        stderr=subprocess.PIPE,
+        stdout=subprocess.PIPE,
+        encoding='UTF-8'
+    )
+    name_in_db = utils.get_run_data(run_name)['name']
+    assert name_in_db == run_name, Tests.assertion_info(f"name: {run_name}", name_in_db)
+
+# --filename FILENAME
+#    An optional alternative filename if you do not want to use "usage_scenario.yml"
+    # basic positive case
+def test_different_filename():
+    run_name = 'test_' + utils.randomword(12)
+    ps = subprocess.run(
+        ['python3', '../runner.py', '--name', run_name, '--uri', GMT_DIR , '--filename', 'tests/data/usage_scenarios/basic_stress.yml', '--config-override', f"{os.path.dirname(os.path.realpath(__file__))}/test-config.yml",
+        '--skip-system-checks', '--dev-no-metrics', '--dev-no-phase-stats', '--dev-no-optimizations', '--dev-no-sleeps', '--dev-cache-build'],
+        check=True,
+        stderr=subprocess.PIPE,
+        stdout=subprocess.PIPE,
+        encoding='UTF-8'
+    )
+
+    with open('data/usage_scenarios/basic_stress.yml', 'r', encoding='utf-8') as f:
+        usage_scenario_contents = yaml.safe_load(f)
+    usage_scenario_in_db = utils.get_run_data(run_name)['usage_scenario']
+    assert usage_scenario_in_db == usage_scenario_contents,\
+        Tests.assertion_info(usage_scenario_contents, usage_scenario_in_db)
+    assert ps.stderr == '', Tests.assertion_info('no errors', ps.stderr)
+
+# if that filename is missing...
+def test_different_filename_missing():
+    runner = ScenarioRunner(uri=GMT_DIR, uri_type='folder', filename='I_do_not_exist.yml', skip_system_checks=True, dev_cache_build=True, dev_no_sleeps=True, dev_no_metrics=True, dev_no_phase_stats=True)
+
+    with pytest.raises(FileNotFoundError) as e:
+        runner.run()
+
+    # we cannot use == here as file paths will differ throughout systems
+    expected_exception = "[Errno 2] No such file or directory"
+    assert expected_exception in str(e.value),\
+        Tests.assertion_info(f"Exception: {expected_exception}", str(e.value))
+
+    expected_exception_2 = "I_do_not_exist.yml"
+    assert expected_exception_2 in str(e.value),\
+        Tests.assertion_info(f"Exception: {expected_exception_2}", str(e.value))
+
+
+#   Check that default is to leave the files
+def test_no_file_cleanup():
+    runner = ScenarioRunner(uri=GMT_DIR, uri_type='folder', filename='tests/data/usage_scenarios/basic_stress.yml', skip_system_checks=True, dev_cache_build=True, dev_no_sleeps=True, dev_no_metrics=True, dev_no_phase_stats=True)
+    runner.run()
+
+    assert os.path.exists('/tmp/green-metrics-tool'), \
+        Tests.assertion_info('tmp directory exists', os.path.exists('/tmp/green-metrics-tool'))
+
+#   Check that the temp dir is deleted when using --file-cleanup
+#   This option exists only in CLI mode
+def test_file_cleanup():
+    subprocess.run(
+        ['python3', '../runner.py', '--uri', GMT_DIR , '--filename', 'tests/data/usage_scenarios/basic_stress.yml',
+         '--file-cleanup', '--config-override', f"{os.path.dirname(os.path.realpath(__file__))}/test-config.yml", '--skip-system-checks', '--dev-no-sleeps', '--dev-cache-build', '--dev-no-metrics', '--dev-no-phase-stats', '--dev-no-optimizations'],
+        check=True,
+        stderr=subprocess.PIPE,
+        stdout=subprocess.PIPE,
+        encoding='UTF-8'
+    )
+    assert not os.path.exists('/tmp/green-metrics-tool'), \
+        Tests.assertion_info('tmp directory exists', not os.path.exists('/tmp/green-metrics-tool'))
+
+#pylint: disable=unused-variable
+def test_skip_and_allow_unsafe_both_true():
+
+    with pytest.raises(RuntimeError) as e:
+        ScenarioRunner(uri=GMT_DIR, uri_type='folder', filename='basic_stress.yml', skip_system_checks=True, dev_cache_build=True, dev_no_sleeps=True, dev_no_metrics=True, dev_no_phase_stats=True, skip_unsafe=True, allow_unsafe=True)
+    expected_exception = 'Cannot specify both --skip-unsafe and --allow-unsafe'
+    assert str(e.value) == expected_exception, Tests.assertion_info('', str(e.value))
+
+def test_debug(monkeypatch):
+    monkeypatch.setattr('sys.stdin', io.StringIO('Enter'))
+    ps = subprocess.run(
+        ['python3', '../runner.py', '--uri', GMT_DIR , '--filename', 'tests/data/usage_scenarios/basic_stress.yml',
+         '--debug',
+         '--config-override', f"{os.path.dirname(os.path.realpath(__file__))}/test-config.yml", '--skip-system-checks',
+          '--dev-no-sleeps', '--dev-cache-build', '--dev-no-metrics', '--dev-no-phase-stats', '--dev-no-optimizations'],
+        check=True,
+        stderr=subprocess.PIPE,
+        stdout=subprocess.PIPE,
+        encoding='UTF-8'
+    )
+    expected_output = 'Initial load complete. Waiting to start metric providers'
+    assert expected_output in ps.stdout, \
+        Tests.assertion_info(expected_output, 'no/different output')
+
+
 
 test_data = [
    (True, f"{os.path.dirname(os.path.realpath(__file__))}/test-config.yml", does_not_raise()),
@@ -273,3 +489,46 @@ def test_runner_filename_with_remote_uri():
     assert ps.returncode == 0
     assert 'Running:  stress/usage_scenario.yml' in ps.stdout
     assert ps.stderr == '', Tests.assertion_info('no errors', ps.stderr)
+
+    ## rethink this one
+def wip_test_verbose_provider_boot():
+    run_name = 'test_' + utils.randomword(12)
+    ps = subprocess.run(
+        ['python3', '../runner.py', '--name', run_name, '--uri', GMT_DIR ,
+         '--verbose-provider-boot', '--config-override', f"{os.path.dirname(os.path.realpath(__file__))}/test-config.yml",
+         '--filename', 'tests/data/stress-application/usage_scenario.yml',
+         '--dev-no-sleeps', '--dev-cache-build', '--dev-no-metrics', '--dev-no-phase-stats', '--dev-no-optimizations'],
+        check=True,
+        stderr=subprocess.PIPE,
+        stdout=subprocess.PIPE,
+        encoding='UTF-8'
+    )
+    run_id = utils.get_run_data(run_name)['id']
+    query = """
+            SELECT
+                time, note
+            FROM
+                notes
+            WHERE
+                run_id = %s
+                AND note LIKE %s
+            ORDER BY
+                time
+            """
+
+    # providers are not started at the same time, but with 2 second delay
+    # there is a note added when it starts "Booting {metric_provider}"
+    # can check for this note in the DB and the notes are about 2s apart
+    notes = DB().fetch_all(query, (run_id,'Booting%',))
+    metric_providers = utils.get_metric_providers_names(GlobalConfig().config)
+
+    #for each metric provider, assert there is an an entry in notes
+    for provider in metric_providers:
+        assert any(provider in note for _, note in notes), \
+            Tests.assertion_info(f"note: 'Booting {provider}'", f"notes: {notes}")
+
+    #check that each timestamp in notes roughly 10 seconds apart
+    for i in range(len(notes)-1):
+        diff = (notes[i+1][0] - notes[i][0])/1000000
+        assert 9.9 <= diff <= 10.1, \
+            Tests.assertion_info('10s apart', f"time difference of notes: {diff}s")
